Add color support for cells
Currently super slow since `Cell` are readonly structs and a lot of
copying is going on that isn't really necessary.

We should change Cell to be a reference type, and update the properties
directly on each cell instead. The cell instances are never really
changing, so this would be a one-time allocation.